Extract cleanUrl and add tests
